Remove obsolete timing helper
